smartschedule/requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
geonamescache==2.0.0
inflection==0.5.1
install==1.3.5
Levenshtein==0.23.0
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
